requirements.txt
fastapi
uvicorn
openai
httpx
pydantic
python-multipart
duckdb
pandas
numpy
matplotlib
seaborn
beautifulsoup4
lxml
aiofiles
python-dotenv
tabulate
fuzzywuzzy
python-Levenshtein
html5lib
scikit-learn
rapidfuzz
requests
boto3